FR-1452 separate bootfs and partitions steps, re-organize code per sil2100 improvements
