+++ internal/statemachine/testdata/grubenv
@@ -1,0 +1,8 @@
+# GRUB Environment Block
+snap_menuentry=Ubuntu Core 18 (amd64)
+snap_core=core18_2128.snap
+snap_kernel=pc-kernel_817.snap
+snap_mode=
+snap_try_core=
+snap_try_kernel=
+############################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################################
